refactor: components/index.ts를 삭제하고, 컴포넌트를 가져올 때 그 컴포넌트의 모듈에서 직접 가져오도록 함 #195
