--- front/src/components/index.ts
@@ -1,20 +1,0 @@
-export { default as TextInput } from './TextInput';
-export { default as SearchBar } from './SearchBar';
-export { default as MakeWorldcupForm } from './MakeWorldcupForm';
-export { default as ImgPreView } from './ImgPreView';
-export { default as ImgPreViewList } from './ImgPreViewList';
-export { default as Header } from './Header';
-export { default as ImgTable } from './ImgTable';
-export { default as TabBar } from './TabBar';
-export { default as StoreBtns } from './StoreBtns';
-export { default as RankingList } from './RankingList';
-export { default as RankingModal } from './RankingModal';
-export { default as Comment } from './Comment';
-export { default as PrivateRoute } from './PrivateRoute';
-export { default as PublicRoute } from './PublicRoute';
-export { default as Loader } from './Loader';
-export { default as RoundSelector } from './RoundSelector';
-export { default as SocialLoginButton } from './SocialLoginButton';
-export { default as Keywords } from './Keywords';
-export { default as WorldcupList } from './WorldcupList';
-export { default as UserInfoForm } from './UserInfoForm';
