+++ front/src/pages/MyInfo/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import { Link } from 'react-router-dom';
-import { Header } from '../../components';
+import Header from '../../components/Header';
 import logo from '../../images/logo.png';
 import { PROFILE, LEAVE } from '../../constants/route';
